src/qumo/optimizers/quantum_optimizer.py
"""QUMO-based optimizer for linux-wallpaperengine-gtk."""
from typing import Dict, List, Optional, Set, Tuple, Any
import ast
import libcst as cst
from sympy import Symbol
import networkx as nx

class QuantumWallpaperTransformer(cst.CSTTransformer):
    """Quantum-inspired code transformer for wallpaper engine."""

    def __init__(self) -> None:
        super().__init__()
        self.ψ = Symbol('ψ')
        self.ɢ = nx.DiGraph()
        self.ᴡ = {}
        self.ᴘᴀᴛᴛᴇʀɴꜱ = {'self': 'ꜱ', 'path': 'ᴘ', 'window': 'ᴡ', 'list': 'ʟ', 'next': 'ɴ', 'prev': 'ᴘ', 'random': 'ʀ', 'current': 'ᴄ', 'display': 'ᴅ', 'engine': 'ᴇ', 'load': 'ʟ', 'save': 'ꜱ', 'update': 'ᴜ', 'validate': 'ᴠ', 'entangle': '⊗', 'superpose': '⊕', 'measure': '⟨·⟩'}
        self.ǫᴜᴀɴᴛᴜᴍ_ᴏᴘꜱ = {'compose': '∘', 'tensor': '⊗', 'direct_sum': '⊕', 'conjugate': '†'}

    def optimize_name(self, name: str) -> str:
        """Convert names to quantum-optimized patterns."""
        return self.ᴘᴀᴛᴛᴇʀɴꜱ.get(name, name)

    def apply_quantum_pattern(self, node: cst.CSTNode) -> cst.CSTNode:
        """Apply quantum-inspired code patterns."""
        match node:
            case cst.Name(value=v):
                return cst.Name(value=self.optimize_name(v))
            case cst.Call(func=f):
                if isinstance(f, cst.Name):
                    op = self.ǫᴜᴀɴᴛᴜᴍ_ᴏᴘꜱ.get(f.value)
                    if op:
                        return self.transform_to_quantum_op(node, op)
        return node

    def transform_to_quantum_op(self, node: cst.Call, op: str) -> cst.Call:
        """Transform function calls to quantum operations."""
        match op:
            case '⊗':
                return self.build_tensor_product(node)
            case '⊕':
                return self.build_direct_sum(node)
            case '†':
                return self.build_conjugate(node)
            case _:
                return node

    def build_tensor_product(self, node: cst.Call) -> cst.Call:
        """Build a quantum tensor product for composed operations."""
        return cst.Call(func=cst.Name('compose_operations'), args=[cst.Arg(value=arg.value) for arg in node.args])

    def build_direct_sum(self, node: cst.Call) -> cst.Call:
        """Build a quantum direct sum for parallel operations."""
        return cst.Call(func=cst.Name('parallel_operations'), args=[cst.Arg(value=arg.value) for arg in node.args])

    def build_conjugate(self, node: cst.Call) -> cst.Call:
        """Build a quantum conjugate for reversible operations."""
        return cst.Call(func=cst.Name('reverse_operation'), args=[cst.Arg(value=arg.value) for arg in node.args])

    def visit_Name(self, node: cst.Name) -> cst.Name:
        """Visit and transform name nodes."""
        return self.apply_quantum_pattern(node)

    def visit_Call(self, node: cst.Call) -> cst.Call:
        """Visit and transform function calls."""
        return self.apply_quantum_pattern(node)

class FlowGraphOptimizer:
    """Graph-based flow optimizer for wallpaper engine."""

    def __init__(self) -> None:
        self.ɢ = nx.DiGraph()
        self.ᴄ = {}

    def analyze_flow(self, ast_tree: ast.AST) -> None:
        """Build and analyze the control flow graph."""
        for node in ast.walk(ast_tree):
            if isinstance(node, (ast.If, ast.While, ast.For)):
                self.ɢ.add_edge(node, node.body[0])
                if hasattr(node, 'orelse') and node.orelse:
                    self.ɢ.add_edge(node, node.orelse[0])

    def optimize_flow(self, ast_tree: ast.AST) -> ast.AST:
        """Optimize code flow using graph analysis."""
        self.analyze_flow(ast_tree)
        try:
            critical_paths = list(nx.all_simple_paths(self.ɢ, source=next((n for n in self.ɢ.nodes() if self.ɢ.in_degree(n) == 0)), target=next((n for n in self.ɢ.nodes() if self.ɢ.out_degree(n) == 0))))
            for path in critical_paths:
                for node in path:
                    if isinstance(node, ast.Call):
                        self.ᴄ[node.func.id] = True
        except (StopIteration, nx.NetworkXNoPath):
            pass
        return ast_tree

def optimize_wallpaper_code(source: str) -> str:
    """
    Apply QUMO optimization to wallpaper engine code.
    
    Args:
        source: Source code to optimize
    
    Returns:
        Optimized source code with quantum patterns
    """
    source_tree = cst.parse_module(source)
    quantum_transformer = QuantumWallpaperTransformer()
    transformed_tree = source_tree.visit(quantum_transformer)
    flow_optimizer = FlowGraphOptimizer()
    ast_tree = ast.parse(transformed_tree.code)
    optimized_ast = flow_optimizer.optimize_flow(ast_tree)
    return ast.unparse(optimized_ast)